Add files via upload
A movie rating example which collects the data of users' rating on different movies. By using the Pearson similarity to find the user similarity and item similarity. After filtering and using the prediction formula, we can get the rating prediction for users in all movies. Then choose the top 20 movies which have high predict ratings and recommend them to the user.